remove old homepage and rename homepage
