+++ packages/2018/src/components/HomePage/PlatformExplorer.js
